+++ ui/backend/src/k8s/oauth.ts
@@ -8,8 +8,8 @@ import { deleteCookie } from './cookies';
 import { jsonRequest } from './json-request';
 import { getToken, K8S_ACCESS_TOKEN_COOKIE } from './token';
 import { redirect, respondInternalServerError, unauthorized } from './respond';
-import { OAUTH_ROUTE_PREFIX, ZTPFW_UI_ROUTE_PREFIX } from '../constants';
 import { setDead } from '../endpoints';
+import { OAUTH_ROUTE_PREFIX, ZTPFW_UI_ROUTE_PREFIX } from '../common';
 
 const logger = console;
 
@@ -146,8 +146,8 @@ export async function logout(req: Request, res: Response): Promise<void> {
 
   const host = req.headers.host;
 
-  deleteCookie(res, { cookie: 'connect.sid' });
   deleteCookie(res, { cookie: K8S_ACCESS_TOKEN_COOKIE });
+  deleteCookie(res, { cookie: 'connect.sid' });
   deleteCookie(res, { cookie: '_oauth_proxy', domain: `.${host || ''}` });
   res.writeHead(200).end();
 }